requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anyio==3.7.1
attrs==23.1.0
auto-gptq==0.2.0
backoff==2.2.1
bitsandbytes==0.41.3.post2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
cohere==4.39
contourpy==1.2.1
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.14
datasets==2.15.0
dill==0.3.7
diskcache==5.6.3
fastapi==0.105.0
fastavro==1.9.2
filelock==3.13.1
Flask==3.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
gpt4all==2.0.2
h11==0.14.0
h3==3.7.7
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jplephem==2.22
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.149
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.71
llama-cpp-python==0.1.49
llama2==0.0.1.dev0
llvmlite==0.41.1
loguru==0.7.2
loralib==0.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nomic==2.0.16
numba==0.58.1
numexpr==2.8.8
numpy==1.26.2
openai-whisper==20231117
openapi-schema-pydantic==1.2.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
peft==0.7.1
Pillow==10.1.0
playsound==1.3.0
plotly==5.18.0
proto-plus==1.24.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.14.5
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pygpt4all==1.1.0
pygptj==2.0.3
pyllamacpp==2.4.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
replicate==0.22.0
requests==2.31.0
retrying==1.3.4
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.2
rsa==4.9
rubicon-objc==0.4.7
safetensors==0.4.1
scipy==1.11.4
selenium==4.16.0
sentencepiece==0.1.99
sgp4==2.23
shapely==2.0.5
six==1.16.0
skyfield==1.49
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
tiktoken==0.5.2
timezonefinder==6.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.1
trio==0.23.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
tzwhere==3.0.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
validators==0.22.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wonderwords==2.2.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
